+++ src/autobahn_wrapper.hpp
@@ -17,9 +17,8 @@
 #pragma clang diagnostic ignored "-Wignored-qualifiers"
 #pragma clang diagnostic ignored "-Wnon-virtual-dtor"
 #pragma clang diagnostic ignored "-Wunused-parameter"
-#if !defined __APPLE__ && __clang_major__ >= 11
+#pragma clang diagnostic ignored "-Wnull-pointer-subtraction"
 #pragma clang diagnostic ignored "-Wdeprecated-copy"
-#endif
 #endif
 
 #pragma GCC diagnostic push