+++ tools/buildgdk_rust.sh
@@ -70,8 +70,12 @@ elif [ \( "$1" = "--iphone" \) ]; then
 elif [ \( "$1" = "--iphonesim" \) ]; then
     RUSTTARGET=x86_64-apple-ios
     LD_ARCH="-arch x86_64 -platform_version ios-simulator 11.0 11.0"
+elif [ "$(uname)" = "Darwin" ]; then
+    SDK_CPU=$(uname -m)
+    RUSTTARGET=$HOST_OS
+    LD_ARCH="-arch $SDK_CPU -platform_version macos 10.13 10.13"
 else
-    LD_ARCH="-arch x86_64 -platform_version macos 10.13 10.13"
+    LD_ARCH="-platform_version macos 10.13 10.13"
 fi
 
 if [ "$BUILDTYPE" == "release" ]; then