+++ tools/buildopenssl.sh
@@ -74,7 +74,8 @@ elif [ \( "$1" = "--windows" \) ]; then
     make install_sw
 else
     if [ "$(uname)" = "Darwin" ]; then
-        ./Configure darwin64-x86_64-cc --prefix="$openssl_prefix" $OPENSSL_OPTIONS -mmacosx-version-min=10.13
+        ARCH=$(uname -m)
+        ./Configure darwin64-$ARCH-cc --prefix="$openssl_prefix" $OPENSSL_OPTIONS -mmacosx-version-min=10.13
     else
         ./config --prefix="$openssl_prefix" $OPENSSL_OPTIONS
         $SED -ie "s!^CFLAG=!CFLAG=-fPIC -DPIC !" "Makefile"